+++ sqldelight-gradle-plugin/src/main/kotlin/app/cash/sqldelight/gradle/PropertiesImpl.kt
@@ -26,6 +26,7 @@ data class SqlDelightDatabasePropertiesImpl(
   @Input override val treatNullAsUnknownForEquality: Boolean = false,
   @Input override val generateAsync: Boolean = false,
   @Input override val expandSelectStar: Boolean = true,
+  @Input override val codegenExcludedColumns: Set<String> = emptySet(),
   // Only used by intellij plugin to help with resolution.
   @Internal override val rootDirectory: File,
 ) : SqlDelightDatabaseProperties