+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [PostgreSQL Dialect] Add support for ALTER INDEX (#6224 by @griffio)
 - [SQLite Dialect] Add support for SQLite 3.44 aggregate functions DISTINCT, ORDER BY and FILTER (#6236 by @griffio)
 - [SQLite Dialect] Add support for SQLite 3.37 STRICT table (#6230 by @griffio)
+- [Gradle Plugin] Add support for excluding columns from generated models with `codegenExcludedColumns` (#6243 by @sokolikp)
 - [Compiler] Add `allTableNames` function to schema (#6245 by @edenman)
 
 ### Changed